+++ frontend/src/components/common/chatting/RoomEditor.tsx
@@ -2,10 +2,10 @@ import React, { useState, useRef, useEffect } from "react";
 import { useDispatch } from "react-redux";
 import { CurrentChattingActionTypes } from "types/redux/CurrentChatting";
 import { ChattingRoomForm } from "types/Form";
-import CustomInput from "components/common/utils/CustomInput";
-import CustomSlider from "components/common/utils/CustomSlider";
-import ChattingTypeSelect from "components/common/utils/ChattingTypeSelect";
-import CustomIconButton from "components/common/utils/CustomIconButton";
+import CustomInput from "components/utils/CustomInput";
+import CustomSlider from "components/utils/CustomSlider";
+import ChattingTypeSelect from "components/utils/ChattingTypeSelect";
+import CustomIconButton from "components/utils/CustomIconButton";
 import ChattingService from "API/ChattingService";
 import { useSelector } from "react-redux";
 import { IRootState } from "components/common/store";
