+++ frontend/src/components/profile/GameHistoryList.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link } from "react-router-dom";
 import { History } from "types/Profile";
-import EmptyListMessage from "components/common/utils/EmptyListMessage";
+import EmptyListMessage from "components/utils/EmptyListMessage";
 import "styles/global.scss";
 import "styles/Profile.scss";
