+++ frontend/src/components/utils/CustomSlider.tsx
