+++ frontend/src/pages/Profile.tsx
@@ -8,8 +8,8 @@ import AchievementList from "components/profile/AchievementList";
 import EditNicknameModal from "components/profile/EditNicknameModal";
 import EditImageModal from "components/profile/EditImageModal";
 import SetTwoFactorModal from "components/profile/SetTwoFactorModal";
-import SkeletonProfile from "components/common/utils/SkeletonProfile";
-import SkeletonButtons from "components/common/utils/SkeletonButtons";
+import SkeletonProfile from "components/utils/SkeletonProfile";
+import SkeletonButtons from "components/utils/SkeletonButtons";
 import MyButtons from "components/profile/MyButtons";
 import OthersButtons from "components/profile/OthersButtons";
 import { MyInfoActionTypes } from "types/redux/MyInfo";
